requirements.txt
python-telegram-bot==20.6
gspread==6.0.0
google-auth==2.26.1
